requirements.txt
opencv-python>=4.8.1.78
onnxruntime>=1.15.0
numpy<2.0.0
Pillow>=8.0.0
mtcnn-runtime>=0.1.0
requests==2.31.0 